requirements.txt
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
boto3==1.34.84
python-dotenv==1.0.1
streamlit==1.11.0
